// src/redux/messages/messages.types.ts
export enum MessagesActionTypes {
  SEND_MESSAGE_LOCAL = 'messages/SEND_MESSAGE_LOCAL',
  SEND_MESSAGE_START = 'messages/SEND_MESSAGE_START',
  SEND_MESSAGE_FAILURE = 'messages/SEND_MESSAGE_FAILURE',
  SEND_MESSAGE_SUCCESS = 'messages/SEND_MESSAGE_SUCCESS',
  FETCH_LAST_MESSAGES_START = 'messages/FETCH_LAST_MESSAGES_START',
  FETCH_LAST_MESSAGES_FAILURE = 'messages/FETCH_LAST_MESSAGES_FAILURE',
  FETCH_LAST_MESSAGES_SUCCESS = 'messages/FETCH_LAST_MESSAGES_SUCCESS',
  FETCH_DIALOGUE_START = 'messages/FETCH_DIALOGUE_START',
}